Remplate intval() par floatval() sinon les quotas (ex: bande passante) sont limites a 2147483647 (1.99 gigs)

+++ bureau/admin/adm_dodefquotas.php
@@ -85,7 +85,7 @@ if($_POST["action"] == "add") {
       continue;
 
     list($type, $q) = explode(":", $key, 2);
-    $c[$type][$q] = abs(intval($val));
+    $c[$type][$q] = abs(floatval($val));
   }
 
   if($quota->setdefaults($c)) {
